:white_check_mark: Improve test coverage and logging

+++ tests/test_yt2mp3.py
@@ -8,7 +8,7 @@ yt2mp3/tests/test_yt2mp3.py
 Brett Stevenson (c) 2018
 """
 
-import os, pytest, yt2mp3, shutil
+import os, pytest
 from mutagen.id3 import ID3
 from collections import defaultdict
 from yt2mp3 import util, opts
@@ -20,7 +20,7 @@ def test_data():
              'collection_name': 'Experience Hendrix: The Best of Jimi Hendrix',
              'primary_genre_name': 'Rock',
              'track_name': 'Bold As Love',
-             'track_number': 12,
+             'track_number': 12
             }
 
 @pytest.fixture
@@ -32,9 +32,7 @@ def test_song():
 def test_arguments():
     errors = []
     args = opts.parseOptions(['-o', '-v', '-a', 'jimi', 'hendrix', '-t', 'bold', 'as', 'love'])
-    assert args.overwrite
-    assert args.verbose
-    assert not args.quiet
+    assert args.overwrite and args.verbose and not args.quiet
     assert args.resolution == 480
     assert ' '.join(args.artist) == 'jimi hendrix'
     assert ' '.join(args.track) == 'bold as love'
@@ -50,7 +48,6 @@ def test_get_song_data(test_data):
     data = defaultdict(str, util.getSongData(input))
     assert [test_data[key] == data[key] for key in test_data.keys()]
 
-
 def test_itunes_data(test_data):    
     data, results = defaultdict(str), dict()
     data['track_name'] = 'bold as love'
@@ -60,17 +57,27 @@ def test_itunes_data(test_data):
     results['artist_name'] = util.getSongData(data)
     assert [len(results[key]) > 2 for key in results.keys()]
 
-def test_validate_URL():
-    urls = ['https://www.youtube.com/watch?v=gkJhnDkdC-0', 'http://youtu.be/gkJhnDkdC-0']
-    playlists = ['https://www.youtube.com/playlist?list=PLGqB3S8f_uiLkCQziivGYI3zNtLJvfUWm', 'https://www.youtube.com/watch?v=gkJhnDkdC-0&list=PLGqB3S8f_uiLkCQziivGYI3zNtLJvfUWm', 'https://www.youtube.com/watch?list=PL1F9CA2A03CF286C2&v=pFS4zYWxzNA&']
-    for url in urls:
-        assert util.validateURL(url)
-    for url in playlists:
-        assert util.validateURL(url, True)
+# Test iTunes lookup failure => Program should exit 
+def test_itunes_data_failure():
+    with pytest.raises(SystemExit) as err:
+        util.getiTunesData('nomatch', 'test')
+    assert err.type == SystemExit
 
-def test_get_video_URL():
-    url = util.getVideoURL('Bold as Love', 'Jimi Hendrix')
-    assert len(url.split('watch?v=')) == 2
+def test_get_video_URL(test_song):
+    url = util.getVideoURL(test_song.track, test_song.artist)
+    assert url == test_song.video_url and util.validateURL(url)
+
+def test_validate_URL():
+    errors = []
+    urls = ['https://www.youtube.com/watch?v=gkJhnDkdC-0', 'http://youtu.be/gkJhnDkdC-0']
+    playlists = ['https://www.youtube.com/playlist?list=PLGqB3S8f_uiLkCQziivGYI3zNtLJvfUWm','https://www.youtube.com/watch?v=gkJhnDkdC-0&list=PLGqB3S8f_uiLkCQziivGYI3zNtLJvfUWm', 'https://www.youtube.com/watch?list=PL1F9CA2A03CF286C2&v=pFS4zYWxzNA&']
+    for url in urls:
+        if not util.validateURL(url):
+            errors.append('Failed URL Validation: '+url) 
+    for url in playlists:
+        if not util.validateURL(url, playlist=True):
+            errors.append('URL Validation: '+url)
+    assert not errors, 'errors occured:\n{}'.format('\n'.join(errors))
 
 def test_video_data(test_data):
     url = 'https://www.youtube.com/watch?v=gkJhnDkdC-0'
@@ -94,12 +101,7 @@ def test_convert_mp3(test_song):
     song_path = test_song.convertToMP3(video_path)
     if not os.path.exists(song_path):
         errors.append('The output MP3 file doesn\'t exist')
-    if os.path.exists(video_path):
-        errors.append('The video file wasn\'t deleted after conversion')
     assert not errors, 'errors occured:\n{}'.format('\n'.join(errors))
-
-def test_file_check(test_song):
-    assert test_song.fileExists()
 
 def test_set_id3_tags(test_song):
     errors = []
@@ -122,28 +124,21 @@ def test_get_playlist_videos():
     videos = util.getVideoList(url)
     assert videos == ['https://www.youtube.com/watch?v=gkJhnDkdC-0', 'https://www.youtube.com/watch?v=_FrOQC-zEog', 'https://www.youtube.com/watch?v=yvPr9YV7-Xw']
 
-def test_lookup_failure():
-    with pytest.raises(SystemExit) as err:
-        util.getiTunesData('nomatch', 'test')
-    assert err.type == SystemExit
-
-def test_overwrite(test_song):
-    song_path = os.path.join(os.path.expanduser('~/Downloads/Music/'), test_song.artist, test_song.track+'.mp3')
-    assert os.path.exists(song_path)
-
+# Make sure temporary files are being cleaned up
 def test_cleanup(test_song):
     errors = []
     directory = os.path.expanduser('~/Downloads/Music/')
     video_dir = os.path.join(directory, 'temp/')
     cover_dir = os.path.join(directory, 'CoverArt/')
     song_path = os.path.join(directory, test_song.artist, test_song.track+'.mp3')
-    os.remove(song_path)
-    if len(os.listdir(os.path.dirname(song_path))) < 1:
-        shutil.rmtree(os.path.dirname(song_path))
+    util.cleanup()
+    # Remove test mp3 file and artist diectory if empty
+    if test_song.fileExists():
+        os.remove(song_path)
+    if os.listdir(os.path.dirname(song_path)):
+        os.rmdir(os.path.dirname(song_path))
     if os.path.isdir(video_dir):
-        errors.append('The video file wasn\'t deleted after conversion')
-        shutil.rmtree(video_dir)
+        errors.append('The temporary video files weren\'t cleaned up')
     if os.path.isdir(cover_dir):
-        errors.append('The cover image wasn\'t deleted after embedding')
-        shutil.rmtree(cover_dir)
+        errors.append('The temporary cover-art files weren\'t cleaned up')
     assert not errors, 'errors occured:\n{}'.format('\n'.join(errors))
